+++ src/ObjectOrientedProgramming/Inheritance/Employee.java
@@ -3,7 +3,7 @@ package ObjectOrientedProgramming.Inheritance;
 public class Employee {
     String firstName;
     String lastName;
-    int salary;
+    double salary;
 
     public Employee(){
         this.firstName = "";
@@ -11,7 +11,7 @@ public class Employee {
         this.salary = 0;
     }
 
-    public Employee(String firstName, String lastName, int salary){
+    public Employee(String firstName, String lastName, double salary){
         this.firstName = firstName;
         this.lastName = lastName;
         this.salary = salary;
